requirements.txt
django
django-bakery
Pillow
sorl-thumbnail
django-bootstrap4
django-localflavor
pdfkit
